Fix restoring from live medium.
When data is stored on the live medium on the same partition as the live
system the partition is already mounted when redo tries to mount it on
/mnt/backup.

It sometimes just works and sometimes mount complains the partition is
busy. If mounting fails attempt to locate a mountpoint where the
partition is already mounted and bind-mount it.