Removes bash script used for creating tarball source code release snapshots
Removes the bash script that appears to have been used for creating the source
code tarballs uploaded to Sourceforge in 2011/2012. The script is obsolete: now
that source code is managed in version control, releasing a new version means
creating a git tag and pushing it publicly.

It's unclear why the snapshot.list contains a subset of the files in each
tarball.