Switches wallpaper app from nitrogen to pcmanfm
Removes the 'nitrogen' wallpaper app, and replaces its functionality with PCMan
File Manager launched in desktop mode, along with a configuration file.  Unlike
nitrogen, pcmanfm is able displays *.desktop files in ~/Desktop/ as desktop
icons, in addition to being able to display wallpapers.  pcmanfm is already
installed on the system, so no need for nitrogen's ~0.5MB deb package.

Adds a pcmanfm configuration file pointing to the wallpaper [1]. The
application centers the image, and uses the 'desktop_bg' to configure the
region outside the image. desktop_bg was set to hex color #233750, based on
examining [1] in an image editor.

[1] /usr/share/backgrounds/redo_blue.png