Switches start-menu/taskbar/systray app to lxpanel
Removes the abandoned adeskbar taskbar application [1], and replaces its
functionality with the LXDE project's lxpanel, an application which continues
to be maintained (and is available in the Ubuntu repositories).

Also launches the systray NetworkManager applet (nm-applet) using openbox
autostart.sh, as nm-applet does not start without this (according to `ps aux`).
It's unclear how adeskbar was launching nm-applet.

Configuration of lxpanel to imitate adeskbar's appearance will occur in a
subsequent commit.

adeskbar contributed enormously to the existing look and feel of Redo Backup
and Recovery (since its addition in June 2011's v0.9.8). Switching to an
alternative launcher (such as lxpanel) is the simplest and most robust path for
future maintenance of Redo Backup and Recovery, however with the desire to
build a Redo Backup and Recovery ISO image based on Ubuntu 18.04 with as few
changes as possible, an attempt was made to install latest adeskbar release
(v0.5.0) on Ubuntu 18.04 (Bionic Beaver). This attempt initially failed due to
missing dependencies: python-wnck, gnome-menus and python-gmenu.  After some
dependencies such as python-wnck, were satisfied by downloading the package
from an older release (end-of-life) release, Ubuntu 17.10 (Artful Aardvark),
adeskbar successfully launched. However once launched, the start menu was
missing. This functionality likely comes from gnome-menus and python-gmenu
packages, which possibly could be replaced with a dependency on mate-menus and
python-mate-menu. Given this overhead, switching from adeskbar to lxpanel was
deemed easier.

[1] https://github.com/adcomp/adeskbar-0.5 (last release Dec 2015)