Modify lxpanel appearance to resemble adeskbar
Modifies lxpanel configuration files to try and match the 'adeskbar'
start-menu/taskbar/systray application used in Redo Backup and Recovery v1.0.4.
adeskbar contributes enormously to the look and feel of Redo Backup and
Recovery (since its addition in June 2011's v0.9.8), and its usability.

Updates configuration files in the following ways:

* Removes "Application Launch Bar" widget
  * These are the icons next the start-menu, what Windows calls "pinned apps"
* Increases taskbar height to 41 pixels, icon size from default 24 pixels to 33
  * Size now matches adeskbar's start menu icon
* Removes 'minimize all windows' button
  * adeskbar did not have an icon to show desktop, nor is one needed
* Removes desktop pager widget
  * Widget is "mini-map" of current workspace
* Removes CPU usage monitor widget
  * Widget provided graph of CPU usage
* Adds battery icon
  * adeskbar did not have it, but may assist laptop users
* Configures digital clock formatting to match adeskbar v1.0.4 clock widget
* Changes start menu icon to adeskbar image

Unfortunately the height of each lxpanel start menu entry is much smaller than
adeskbar and is not configurable yet. This may be an issue on HiDPI displays.