Adds shell script to checksum an ISO image and squashfs root filesystem
Adds a helper bash script to mount an ISO9660 image, extract the compressed
(squashfs) filesytem, and checksum all files on both the ISO9660 and squashfs
filesystems. This checksum file is very useful for comparing an officially
released ISO image to an ISO built from source, especially used the graphical
diff tool 'meld'.

The script will continue to be useful in auditing any third-party ISO image
against the source code it was generated from. The analysis using 'meld'
provides an indication of what differs, which suggest what areas to manually
investigate further. Differences may be attributable to use of backports and
updates APT repositories, which get updated with new versions of packages
during the support window of an Ubuntu release.