Installs build deps (debootstrap etc) in Dockerfile
Moves installation of build dependencies from the build script to the
Dockerfile. The build scripts were based as closely as possible on the Redo
Backup and Recovery Sourceforge Wiki instructions v1.0.4 [1], which installs
packages on the host-system. Modifying the host-system package environment
using build scripts is highly uncommon behavior, so these operations are moved
to the Dockerfile.

[1] https://web.archive.org/web/20190924102811/https://sourceforge.net/p/redobackup/wiki/Home/?version=11