Extracts locale.nopurge from heredoc to text file
Extracts the locale.nopurge from a bash "heredoc" into a standalone text file.
Git history of standalone text files are more easily manage, and more cleanly
represented than a heredoc in the build scripts.

Historically, a heredoc was used to stay as close as possible to the Redo
Backup and Recovery Sourceforge Wiki instructions [1] (which required
interactive operation to configure the localepurge package). However, given the
Wiki instructions are obsolete (and the git managed build scripts are now the
ground truth), there is no longer any need to stay close to the Wiki
instructions.

[1] https://web.archive.org/web/20190924102811/https://sourceforge.net/p/redobackup/wiki/Home/?version=11