Combines pkg install into one `apt-get install`
Calls `apt-get install` once with a long list of packages, instead of 3 times
with 3 shorter lists.

The second approach was how the Redo Backup and Recovery Sourceforge Wiki
instructions [1] did it. A single `apt-get install` invocation has a number of
advantages: the log output is clearer to a developer to read (and scroll past),
and any download progress bars more meaningfully display the remaining packages
to download.

[1] https://web.archive.org/web/20190924102811/https://sourceforge.net/p/redobackup/wiki/Home/?version=11