Customizes lxpanel startmenu to match Redo v1.0.4
Deletes the adeskbar user start menu configuration files and configures the
lxpanel start menu to match Redo Backup and Recovery v1.0.4 [1], as cleanly and
as possible. In theory, given Linux start menu configuration is based on
freedesktop's menu-spec [2], the start menu should be identical between
adeskbar and lxpanel without any changes, but this was not the case -- likely
due to adeskbar not following the specification fully.

The new XML menu was created using a combination of pcmanfm's menu editor (Go
-> Applications), and manual modification modification.

The biggest difference between the adeskbar entries is the use of 'Name' field
(eg. "GParted"), instead of 'GenericName' (eg. "Partition Editor"). This
effects usability for users who aren't familiar with the names of Linux tools.
This difference may be fixed as a subsequent commit.

[1]
Redo Backup & Recovery v1.0.4 Start Menu
Accessories →
		File Manager
		Image Viewer
		Terminal
		Text Editor
		Web Browser
Disk Tools →
		Disk Utility
		Drive Reset
		File Recovery
		Logical Volumes
		Partition Editor
		Usage Analyzer
Settings →
		Create Bootable USB
		Customize Appearance
		Hardware Lister
		Monitor Settings
		Window Settings

[2] https://specifications.freedesktop.org/menu-spec/menu-spec-1.0.html