Fixes pcmanfm 'Terminal emulator is not set' error
Fixes the error that pcmanfm file manager outputs when using Tools -> Open
Current Folder in Terminal, by populating pcmanfm's libfm.conf file. Working
settings are automatically generated inside libfm.conf when it is not present,
so deleted libfm.conf and re-launching pcmanfm.

The only option displayed as modified was 'use_trash', so that was configured
to continue to be the non-default value used prior to this commit.

This issue was present on past versions of Redo Backup and Recovery.