Replaces buggy lxrandr with arandr screen layout
Replaces buggy lxrandr (used to modify screen resolution) with less-buggy, but
less user-friendly arandr utility.

lxrandr was already installed and is a simple, less flexible, less well-known,
but much more user-friendly utility.  However, it has sometimes failed to
modify resolution of the display when VirtualBox, and arandr was able to modify
the resolution successfully in such cases.  Ideally the much more user-friendly
lxrandr would be provided as that's the best user-experience for users
unfamiliar to arandr, but in testing it was too buggy.

Therefore, providing the much more well-known arandr to users.