Adds GNU ddrescue package for dying disk recovery
Adds GNU ddrescue, which is similar to `dd` but handles errors, is able to
resume interrupted operations, and has an retry algorithm to copy all working
blocks before retrying problematic blocks.

ddrescue in a CLI tool meant for advanced users. A subsequent commit may add a
GUI, depending on feedback.