Launches chromium using --no-sandbox, as X is root
Launches chromium browser with the --no-sandbox flag, to allow it to be
launched as root user. Redo Backup and Recovery v1.0.4 also had chromium
present and the X session owned by root user. However the chromium version on
Ubuntu 12.04 did not need the --no-sandbox argument to launch.

The use of --no-sandbox is a workaround, which displays an informational bar
saying "You are using an unsupported command-line flag --no-sandbox. Stability
and security will suffer'.

The correct solution is to launch the X session as a non-root user. However
this will require adding PolicyKit configuration files, and using pkexec to
ensure that all other applications (such as Redo Backup and Recovery itself)
are launched as root. Given the relatively large amount of effort involved in
doing that, the --no-sandbox workaround is a sufficient compromise for the next
Redo Backup and Recovery release.