Adds `git describe`-based version to ISOLINUX menu
Adds a version string to the boot menu, so unmarked boot media can be quickly
identified.

Increases width of maximum ISOLINUX menu text, and shifts horizontal offset
slightly left. This ensures that the output of `git describe --tags` non-tagged
releases displays get adequately represented on the screen with minimal
truncation. Some truncation may still occur, as the output may even be in the
form of X.Y.Z-abc-gGITSHA-dirty, assuming a git tag in the form of 'X.Y.Z', and
a dirty working tree.

Adds a version string to the help text too. This provides the full version
string in a location where truncation will not be an issue.