Increases openbox window sizes for HiDPI displays
Increases the font sizes on openbox for improved usability on HiDPI displays.
This prevents the title bar from being exceptionally small, and the
minimize/maximize/close buttons being difficult to press. The
minimize/maximize/close icons themselves are not scaled, but the clickable box
is.

On low resolutions such as 1024x768, the increased sizes do not hamper
usability.