Adds apt preferences file for backports repository
Lowers the priority of the backports repository to 500, as according to `man
apt_preferences) that selects 'versions that do not belong to the target
release'.

The backports entry in sources.list is provided for long-term maintenance, to
gain access to the features of specific packages when required. However,
backports is always a last resort as it has mere best-effort security
guarantees compared to the main repositories, which are managed by the Ubuntu
Security Team. As the SecurityTeam FAQ [1] suggests: "Packages in release,
security and updates are supported by the Ubuntu Security team, while packages
in backports are supported by the community and packages in proposed are the
responsibility of the uploader."

[1] https://wiki.ubuntu.com/SecurityTeam/FAQ