Sets docker exec working dir instead of make -C
Sets the working directory on the `docker exec` command, so the subsequent
commands take place in the given directory, rather than using `make -C dir` to
"Change to directory dir before reading the makefiles or doing anything else".

This bike-shedding change makes things slightly clearer to somebody who is not
necessarily familiar with software such as GNU Make or docker, but may still
need to run `make install` and `make clean`. An example might be a user wishing
to develop and test a translation.