Renames Makefile 'install' target to 'iso-image'
Renames the first target in the Makefile from 'install' to 'iso-image', and
adds a .DEFAULT_GOAL parameter to explicitly set it as the default target [1]
[2]. Updates the documentation and .travis.yml continuous integration
configuration file to reflect this.

The reason this is important is `make install` is conventionally used for
installing applications on the host system, especially when called with
superuser privileges. However, the build script is not modifying host system,
it requires superuser privileges for build operations such as creating
temporary bind mounts.

[1] "In Make version 3.79 the default target is the first target in the
Makefile."

[2] "In Make 3.81, the default target can be explicitly set using the special
variable .DEFAULT_GOAL e.g."