Documents cleanup before rebuilding docker image
Documents the steps required before the docker image is rebuilt. The docker
build takes place with a UID/GID that is unlikely to be the same as the local
user, which causes an error when the `docker build` process cannot read the
directories containing the existing build artifacts and pkg cache.

When rebuilding the docker image, it makes sense to delete the existing build
artifacts (and the cached pkg indexes). The cache deb files are fixed using
`chown` because they are useful and should not be deleted.