Replaces `git clone` with faster `git lfs clone`
Replaces the `git clone` command with `git lfs clone`, as according to [1]:

Rather than downloading Git LFS files one at a time, the git lfs clone command
waits until the checkout is complete, and then downloads any required Git LFS
files as a batch. This takes advantage of parallelized downloads, and
dramatically reduces the number of HTTP requests and processes spawned (which
is especially important for improving performance on Windows).

[1] https://www.atlassian.com/git/tutorials/git-lfs#clone-respository