Substitutes most 'gzip' strings with 'pigz'
Replaces the invocation of 'gzip' compression utility with the 'pigz'
multithreaded implementation, by running the command [1] from the root of the
git repository.  The CHANGELOG contained a 'gzip' string that was not intended
to be modified, so that file not committed.

The gzip commands are already using -p argument to specify the number of
processes, as the unofficial ISO images released by OAM775 replaced the
/bin/gzip binary itself with the pigz binary. See prior commit for further
information.

[1] git grep -l 'gzip' | xargs sed -i 's/gzip/pigz/g'