Replaces 'mkisofs' command with 'genisoimage'
Replaces the 'mkisofs' call with 'genisoimage'. On Ubuntu 18.04,
/usr/bin/mkisofs is a symlink to /usr/bin/genisoimage. However, on Debian 9
Stretch there is no symlink present, which causes the build to fail if
following the native build instructions (with debootstrap from
stretch-backports, so bionic can be bootstrapped).

Switching to calling genisoimage directly improves quality-of-life for a subset
of users following the native build instructions.