Lengthens ISOLINUX boot menu for more languages
Lengthens the ISOLINUX boot menu, moves the countdown (and '[F1] for help' text
lower), slightly changes the '[F1] for help' text to document '[TAB] to edit
boot option' functionality. Layout documentation can be found here: [1].

These changes are done to be able to add another two languages to the ISOLINUX
boot menu. The 'Simple Menu System' continues to be used, a future version may
switch to the compiled 'Advanced Menu System' for much more flexibility with a
separate menu for language selection. [2]

[1] https://wiki.syslinux.org/wiki/index.php?title=Menu

[2] https://wiki.syslinux.org/wiki/index.php?title=Comboot/menu.c32