Converts text *.ko translations to binary *.mo
Adds to the build script the ability to convert *.ko text files containing
translation strings (which may be generated using xgettext) to the *.mo binary
format used by the GTK API for Glade files. This conversion done as a part of a
script, to remove the need to version-control binary *.mo files.

The conversion requires the msgfmt application, so adds the 'gettext' package
to the host environment (Dockerfile, and build instructions).