Adds French translation for graphical shutdown
Adds a French translation for the graphical shutdown shell script, based on
keywords shared with the lxsession-logout application
(Shutdown/Reboot/Cancel/End Session), and Google translate for the string
''Select an option' ('Choisir une option').

It's unclear how to localize the lxpanel start menu 'Shutdown' menu item
itself. On Lubuntu, localized text is displayed as 'déconnexion'. Searching for
this string in the package lxpanel-common reveals it's part of the *.mo file.
However, the 'Shutdown' string is currently defined (by Redo Backup and
Recovery) in the lxpanel configuration file, because graphical-shutdown.sh is a
custom script.  Ensuring localized 'Shutdown' menu item is left for a
subsequent commit.