Checks error code on key bootloader config steps
Checks the return code on the copying of the vmlinuz image, initrd image and
ISOLINUX bootloader files, and exits if non-zero.

If any of these steps return a non-zero error code, the generated ISO image may
fail to boot completely (for reasons that may be unclear to a developer), so
it's worth exiting the build process with a clear error so that the issue can
be fixed immediately.