Modifies default maximize for several utility apps
Prevents the following utility applications from being auto-maximized on
launch: leafpad text editor, the image viewer, the network connection manager
and the volume mixer.  Disables default-maximize for certain tools like the
Startup Disk Creator, the Hardware Lister, and a loading bar in Redo Backup and
Recovery loading bar.

Also ensures gparted *is* maximized, but the informational windows (with titles
such as 'Applying operations') are not maximized.

This is to increase usability of these applications.