Adds PolicyKit policies to run menu items as root
Adds PolicyKit policy files, pkexec shell wrappers and updates the desktop
entries for applications which need to elevate privileges to root. Currently,
the X session runs as root, but this will change in a subsequent commit.

PolicyKit is the modern 'gksu' privilege request system. Given the 'ubuntu'
user has sudo rights (without a password), no graphical escalation popup is
required.

For simplicity, all custom PolicyKit policy filename are prepended with
'org.redobackup', even when the application vendor is not redobackup.