Deploys home dir only to ubuntu user, not root
Renames /etc/skel to /home/ubuntu, and removes the rsync synchronization of the
/etc/skel directory into the /root directory.

The use of /etc/skel, and synchronizing to /root appears obsolete now that the
X session runs as the ubuntu user.

Removing this duplication reduces the size of the ISO image by 296KB according
to [1], and is likely to make the codebase slightly clearer, as /etc/skel is
usually not the first place a contributor will look.

[1] du -sh src/livecd/chroot/home/ubuntu/