Updates deprecated sfdisk arguments
Removes the use of sfdisk application's '-x' argument, which lists extended
partitions in the output), as the argument has been deprecated according to the
manpage [1].

Replaces the use of sfdisk application's '-R' argument with blockdev, as
suggested by the sfdisk manpage: Since version 2.26 sfdisk no longer provides
the -R or --re-read option to force the kernel to reread the partition table.
Use blockdev --rereadpt instead.

Without this, the version of sfdisk used in Ubuntu 18.04 (2.31.1) an 'invalid
option' message is printed.

For easy reference, the Redo Backup and Recovery 1.0.4 (Ubuntu 12.04) sfdisk
version is 2.20.1.

The extended partition change may effect compatibility with DOS disks [1], but
did not effect the backup and restore use cases that Redo Backup and Recovery
has been manually tested against.

[1] Since version 2.26 sfdisk does not provide the --DOS, --IBM,
--DOS-extended, --unhide, --show-extended, --cylinders, --heads, --sectors,
--inside-outer, --not-inside-outer options.

[2] https://en.wikipedia.org/wiki/Extended_boot_record