Renames /etc/redo_share to /etc/redo/share
Renames the file /etc/redo_share, which contains the default network drive
path, to /etc/redo/share.

The reference to the file was updated in the previous commit (which was
cherry-picked from a git repo on the internet). However, while that commit
updated the reference, it did not rename the file for some reason.