Removes 'values' argument from call to max() fn
Removes the 'values' argument when calling the max() function added in the
previous commit (which was cherry-picked from a git repo on the internet).
Running this using the Perl interpreter bundled with Ubuntu 18.04 produces the
message:

> Experimental values on scalar is now forbidden at /sbin/redobackup line 370.
> Type of arg 1 to values must be hash or array (not subroutine entry) at /sbin/redobackup line 370, near "))"
> Experimental values on scalar is now forbidden at /sbin/redobackup line 783.
> Type of arg 1 to values must be hash or array (not subroutine entry) at /sbin/redobackup line 783, near "))"
> Execution of /sbin/redobackup aborted due to compilation errors.

Looking this up, this was "added in Perl 5.14 but removed in 5.23":
Experimental  %s on scalar is now forbidden (F) An experimental feature added
in Perl 5.14 allowed each, keys, push, pop, shift, splice, unshift, and values
to be called with a scalar argument. This experiment is considered
unsuccessful, and has been removed. The postderef feature may meet your needs
better.

I am not a Perl expert, however looking at the max() function signature, it's
unclear what if the 'values' argument is needed at all. The title of the
previous commit being 'Should allow restoring large drive with small partition
at the start to a small drive' suggests the change was not thoroughly tested
when it was created. Given removing the 'values' argument allows redobackup to
start, and continue to backup and restore, testing this large drive -> small
drive restore functionality (and fixing things if required) will occur as a
future commit.

[1] https://stackoverflow.com/a/41192741