Fixes USB boot functionality via isohybrid package
Fixes the USB boot functionality that was present in Redo Backup and Recovery
v1.0.4, but broke during the upgrade to Ubuntu 18.04 (Bionic). Testing has been
occurring VirtualBox via loading an ISO image into the virtual CD-ROM drive so
this issue was not detected until recently trying to boot on real hardware.