Renames project Rescuezilla in rest of files (2/2)
Renames the entire project Rescuezilla by moving files and directories
containing names like 'redo' or 'redobackup' to instead be named 'rescuezilla'.
Updates references to these paths to ensure everything continues to work.

This was achieved using a series of sed substitutions followed by manual
modifications as required. Certain paths like the (still in use) Redo logo keep
a name which identifies it as belonging to Redo Backup and Recovery.