Exits build if chroot fails to apt-get install
Exits the build script if the main `apt-get install` command in the chroot
fails. Also added an exit code check for the second set of chroot commands that
may be used in subsequent commits.

A typo in a single package name can cause the `apt-get install` command to
fail. Without checking the exit code, the build fails in a subsequent command
for a reason that will not be immediately clear to a developer.