Updates homepage to rescuezilla.com instead of zebradots
Updates the homepage to rescuezilla.com to help provide users with important
up-to-date information and links to the discussion forum. The old homepage was
for 'ZebraDots', which is a website generator side-project by Redo Backup and
Recovery's original author and is otherwise completely unrelated to the Redo
Backup application.

A query-string containing the version number has been appended to help the
Rescuezilla.com GitHub Pages (which is static HTML/JS) [1] to display a banner
when new versions are released.

[1] https://github.com/rescuezilla/rescuezilla.github.io