Ensures Redo Backup 1.0.4 backups can be restored
Ensures backups created by Redo Backup and Recovery v1.0.4 can be restored to
drives given that the versions of partclone.restore and sfdisk on Ubuntu 18.04
have broken backwards compatibility compared to the versions released with
Ubuntu 12.04.  Also writes a version number alongside any new backups, to
remove any ambiguity in the future. Any backup without a version number is
assumed to be v1.0.4.

In Ubuntu 18.04 partclone, restoring with the same tool used to clone (eg,
`partclone.ntfs --restore`) is much much reliable than relying on
auto-detection with `partclone.restore`.  To support this behavior, the
information about which tool was used to create the backup is also written in
the backup directory.

Ensured French and German translations have placeholders by running [1] and
manually extracting the translation placeholders from the diff.

[1] xgettext --no-location --keyword --keyword="loc" --language=Perl \
--output=src/livecd/chroot/usr/share/rescuezilla/translations-rescuezilla/de.po
\ src/livecd/chroot/usr/local/sbin/rescuezilla