Returns manual ellipsizing/truncation to front-end
Returns the manually-ellipsis truncation of long partition names. The removal
of this was imported from an entirely different repository that had a different
author.

After manually importing the commits onto a local branch and testing, it became
clear that Glade does not auto-truncate. The imported truncation commit was
intended to have been removed before the branch was merged but it wasn't, so it
is being removed now.

This reverts commit 33026404eae84034afd745c68295f50f51b55503.