Fixes SMB/CIFS network mount credentials file
Fixes the SMB/CIFS credentials file containing username/password information
that happens to be enclosed in single quotes. Removing the single quotes is
required for correct behavior (the quote special character is not intended to
be part of the password string being passed to mount.cifs).

A prior commit added the single quotes to fix special-character passwords being
broken, and tested behavior successfully. A recent commit imported from a
different repository ensured passwords are not written to stdout by instead
using a credentials file written to /tmp. However, mistakenly, the SMB/CIFS
network mount change was not tested until now.