Displays error msg if unable to umount partition
Displays error message if unable to unmount a partition, and allows the user to
manually close any applications that prevent the umount before retrying.

This prevents the dangerous situation where an application (such as a terminal)
prevents the unmount operation from succeeding, and the *mounted* partition
gets backed up (with the only warning being a very cryptic 'error exit'
message). Backing up a mounted partition may cause a corrupted backup.