Adds link to releases at the bottom of README.md
Adds another link to the GitHub Releases page because GitHub's traffic
analytics show about 15% of traffic visiting the BUILD.ISO.IMAGE.md page.

Given the number of git clones is far less than 15% of the traffic, it appears
non-developer users are scrolling down to the bottom of the page looking for
the download link and clicking the build instructions trying to get do a
download page.