Installs all Xorg video drivers via metapackage
Installs the xserver-xorg-video-all-hwe-18.04 metapackage, rather than just the
generic VESA driver.  This means of the all open source video drivers are again
installed and fixes an issue with v1.0.5 where booting on some hardware fails
to start the graphical environment, leaving users at a text login prompt.

The root cause was that starting from Ubuntu 17.04 (Zesty) the
'xserver-xorg-video-all' metapackage that is used to select all open-source
drivers for installation was been demoted to a recommended package of the
xserver-xorg instead of a dependency. [1] This is why the package was installed
on v1.0.4 (based on Ubuntu 12.04) but not on v1.0.5 (based on Ubuntu 18.04).

[1] https://launchpad.net/ubuntu/+source/xorg/1:7.7+16ubuntu1

Fixes #25