Adds support for backup/restore of NVMe drives
Integrates support for NVMe drive, which have UNIX device nodes like
/dev/nvme0n1p3 (where the nvme0 indicates the first NVMe device, n1 indicates
the first namespace and p3 indicates partition number 3). Previously,
Rescuezilla only supported SATA, IDE/PATA, SAS and SCSI drives, which have UNIX
devices nodes in the form of /dev/sdc4 (where the 'c' corresponds to the third
SATA/SCSI/SAS/USB drive, and the '4' corresponds to the fourth partition).

While the assumption of the /dev/sdXY device node pattern was present in many
places throughout the application structure, the minimal amount of modification
was made to add NVMe support while maintaining the existing application
structure as best as possible.

Rescuezilla now assigns the human-readable drive number ("Drive 1", "Drive 2"
etc) using the output row of the `lsblk` application, rather than its previous
approach of converting the 'd' part of 'sdd' into the number 4. Please note,
the utility `fsarchiver` continues to be used to probe the drives (as it has
always been since the first version of Redo Backup and Recovery).  The
previously unused `lsblk` utility was selected for use here as it is designed
for embedding into scripts and has better column/row filtering options.

Functions have been added to split and rejoin device nodes. This allows the
partition numbers of device nodes to be correctly extracted (for example, 3 in
'nvme0n1p3' or '5' in 'sda5').

Additionally, when iterating over the drive numbers and partition numbers,
numeric sorting is now used (rather than lexical sorting) in all places except
the partitions to save screen. This means instead of "Drive 1, Drive 10, Drive
2, Drive 3,..." it's "Drive 1, Drive 2, Drive 3,..., Drive 10".

Backwards compatibility note: When testing an ISO image build from this commit,
backups made with v1.0.4 and v1.0.5 restore correctly.  All backups created
with an ISO build from this commit restore fine using an v1.0.5 ISO, with the
exception of backups made from NVMe partitions (regardless of the destination
drive technology). This is because the v1.0.5 application processes the source
partition to extract information that's used in the restore process, and v1.0.5
doesn't handle the NVMe device node pattern).

Fixes #27