Removes incorrect MBR 32768 -> 512 byte CHANGELOG message
Removes a misleading line the CHANGELOG of v1.0.5 regarding changing the master
boot record size being copied from 32768 bytes to 512 bytes.

This change was never released in Rescuezilla v1.0.5, however the CHANGELOG
(which was compiled during development) incorrectly states it was.

The change was present in an unofficial Redo Backup and Recovery build so was
integrated during development of v1.0.5 as part of efforts to unify all
versions of Redo Backup and Recovery into a single code base. However, after
testing this change it was removed because it broke booting of operating
systems that were backed up and restored. This breakage is due to GRUB using
the empty 32,256 bytes for its GRUB stage 1.5 boot process.