Stops unmounting source partition on restore
Ensures the source partition (eg, /dev/sdb1) is NOT unmounted during a restore
operation. Unmounting this was mistakenly added in a recent commit. [1]

Given the source partition is read from a file not a block device, the source
partition variable ($part) only indicates what the original source partition
device node was, so is mostly not relevant during the restore process. In fact,
if the sdb1 partition happens to be the source drive where the backups are
stored, this breaks the restore operation. This breakage was not caught during
testing of the prior commit as backup drive and the original partition being
restored happened to not have the same device node string.

[1] 8f4f0dba214eb78f97c45e9ce046b2d9d7223b3d