Writes false *source* partition for NVMe backups
Writes the source partitions in the form of 'sdzC' when backing up NVMe drives,
rather than writing actual source partition (in the form 'nvmeAnBpC'). This
makes a backup created from an NVMe drive using the next Rescuezilla release
(v1.0.5.1) able to successfully be restored to a non-NVMe drive using
Rescuezilla v1.0.5.

The reason for this is Rescuezilla backup and restore operations processes the
source and destination partitions, and v1.0.5 did not support device nodes in
the pattern nvmeAnBpC.

The source partition written to the backup does not really matter: it may have
some usefulness to a end user, but its use in Rescuezilla is just to extract
the each partition number. Therefore to maximize compatibility and restoring
backups with different versions of Rescuezilla as far as practical, the source
partition may as well be written as 'sdzC', rather than the backup file
containing a list of the source partitions in the form 'nvmeAnBpC'.

Please note, the base device 'sdz' has no special significance other than being
the 26th SATA/SCSI/SAS/USB hard drive in the system (which is unlikely to be
present). Also the fact that 'sdz' is unlikely to be present in the system
helps avoid the "accidental unmount of source partition" issue detailed in
issue #36, which has been fixed for the next Rescuezilla release (v1.0.5.1) but
was present in v1.0.5. Again, see issue #36 for full details.

Fixes #27