Lowers X11 text scaling back to 100% not 125%
Lowers the X11 text scaling parameter in the Xresources file because the
PCManFM desktop does not handle it well: at all resolutions, the main
Rescuezilla icon has a line break between the 'Rescuezi' and the final 'la'.

The text scaling is useful for small dimension HiDPI displays, but it does not
seem at all worth it to lowering the polish for all users. If a user of a HiDPI
display experiences issues, they can always configure the Xresources file, or
even easier simply lower the display resolution.