Adds comments about localization of log messages
Adds comments to the Perl script noting that recently added log messages do not
need to be localized by translators. In future existing log messages surrounded
by a loc() localization function call will be removed. This will save
translators work, and means that all the translation work they do is directly
seen by users as intended.

The original French/German translations did translate the log messages, and
that's why Rescuezilla took that approach. However, developers of the Perl
script are expected to be proficient in English, and many endusers will not be
able to understand the log messages even if it was localized into their native
language. Translating log messages is a waste of time and a pain to keep up to
date.

A subsequent commit will remove localizations of existing log messages from the
codebase and from the existing translations files.