Sets default to prevent overwriting source drive
Sets the default behavior of Rescuezilla to provide a message box then exit if
an attempt is made to overwrite the source drive.

v1.0.4 had a confirmation box but there is still a risk that a user will
accidentally click 'yes'. In v1.0.5, several commits were cherry-picked from
Michal Suchanek to check for the existence of [1] which determines whether it's
even possible to overwrite the source. If the file is not present a
confirmation box is still displayed. This commit actually adds this file to
change default behavior.

In theory it can be safe to restore backed up partition to an unmounted
partition on the same device as the source partition. But Rescuezilla does not
yet handle that situation so it's definitely good to completely block all risk.

[1] /etc/rescuezilla/no_source_overwrite