Uses %1 style string interpolation, not [_1]
Switches the string interpolation style from the "[_N]" style used by the
Maketext framework to the "%N" style used by the gettext framework, as some
strings were not being substituted using the maketext style but are correctly
substituted using the gettext style.

All rescuezilla and drivereset localization files were translated using the
following bash script:

for file in "${FILE_LIST[@]}"
do
  for i in $(seq 1 6);
  do
    sed --in-place "s*\[_$i\]*%$i*g" $file
  done
done

drivereset is to be removed in a future release so no testing was done on that
application.