Adds video drivers for Intel and QXL chips
Adds the video driver for Intel chips, as well as QXL. AMD Geode chips not
included as no HWE package present. [1]

An earlier commit [2] intended on adding the Intel video driver by using the
metapackage [3], but the Intel driver was a suggested package. It appears this
wasn't seen because the presence of the driver in the package cache, causing
automatic installation during development builds.

Explicitly listing the suggested packages is the correct solution.

[1] https://packages.ubuntu.com/bionic-updates/xserver-xorg-video-all-hwe-18.04

[2] 2a038ab0c0d76ec4c8145a61e0be0b1b24c80a71

[3] xserver-xorg-video-all-hwe-18.04

Fixes #25