Updates README.md for v1.0.5.1
Adds to the abridged history table information about Rescuezilla v1.0.5.1 and
the old unofficial updates.

Also updates the README.md download link to point to the latest release (that
is, non pre-release tag), rather than simply the link to the GitHub Releases
page.  This should reduce the number of downloads of older versions, which
seems to be due to people scrolling past the latest version.

Finally removes the Patreon link from the bottom of the README.md, to try and
assist users finding the download link.