Switches to Firefox from Chromium to avoid Snaps
Replaces the Chromium web browser with the Firefox web browser, including
updating configuration files, desktop shortcuts, start menu configuration as
relevant. This switch is to avoid the fact that Ubuntu 20.04 (Focal) packages
the Chromium web browser as a snap [1], and snaps cannot be installed in a
chroot environment as of writing [2] (a chroot environment is used by the build
scripts to build the Rescuezilla ISO image).

[1] https://en.wikipedia.org/wiki/Snappy_(package_manager)

[2] https://bugs.launchpad.net/snappy/+bug/1609903